Interview

+++ Algorithm/src/geeks/LargeNumberDivisibleBy8.java
@@ -4,7 +4,7 @@ public class LargeNumberDivisibleBy8 {
 
   private static boolean isDivisibleBy8(String s) {
     s = !s.isEmpty() && s.length() <= 3 ? s : s.substring(s.length() - 3);
-    return Integer.parseInt(s) % 8 == 0;
+    return Long.parseLong(s) % 8 == 0;
   }
 
   public static void main(String[] args) {
